backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
redis[asyncio]==5.0.8
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic-settings==2.3.4
